avoid compiler warning (cast from double to float)

+++ core/vgl/vgl_area.cxx
@@ -1,3 +1,4 @@
+// This is vxl/vgl/vgl_area.cxx
 #include "vgl_area.h"
 #include "vgl_polygon.h"
 
@@ -8,16 +9,16 @@ vgl_area_signed( const vgl_polygon& poly )
 {
   // Compute the area using Green's theorem
   //
-  float area = 0.0;
-  for( int s = 0; s < poly.num_sheets(); ++s ) {
+  float area = 0.0f;
+  for ( int s = 0; s < poly.num_sheets(); ++s ) {
     unsigned int j = poly[s].size()-1;
-    for( unsigned int i = 0; i < poly[s].size(); ++i ) {
+    for ( unsigned int i = 0; i < poly[s].size(); ++i ) {
       area += poly[s][j].x() * poly[s][i].y() - poly[s][i].x() * poly[s][j].y();
       j = i;
     }
   }
 
-  return area/2.0;
+  return area/2;
 }
 
 // This function is not implemented inline because the cost of a
@@ -27,5 +28,5 @@ vgl_area_signed( const vgl_polygon& poly )
 float
 vgl_area( const vgl_polygon& poly )
 {
-  return vcl_fabs( vgl_area_signed( poly ) );
+  return (float)vcl_fabs( vgl_area_signed( poly ) );
 }
